LPS-75989 Move portlet to internal.
