--- modules/apps/foundation/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/portlet/UsersAdminPortlet.java
@@ -1,52 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.users.admin.web.portlet;
-
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.users.admin.constants.UsersAdminPortletKeys;
-
-import javax.portlet.Portlet;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Pei-Jung Lan
- */
-@Component(
-	immediate = true,
-	property = {
-		"com.liferay.portlet.css-class-wrapper=portlet-users-admin",
-		"com.liferay.portlet.display-category=category.hidden",
-		"com.liferay.portlet.header-portlet-css=/css/main.css",
-		"com.liferay.portlet.icon=/icons/users_admin.png",
-		"com.liferay.portlet.preferences-owned-by-group=true",
-		"com.liferay.portlet.private-request-attributes=false",
-		"com.liferay.portlet.private-session-attributes=false",
-		"com.liferay.portlet.render-weight=50",
-		"com.liferay.portlet.struts-path=users_admin",
-		"com.liferay.portlet.use-default-template=true",
-		"javax.portlet.display-name=Users Admin",
-		"javax.portlet.expiration-cache=0",
-		"javax.portlet.init-param.template-path=/",
-		"javax.portlet.init-param.view-template=/view.jsp",
-		"javax.portlet.name=" + UsersAdminPortletKeys.USERS_ADMIN,
-		"javax.portlet.resource-bundle=content.Language",
-		"javax.portlet.security-role-ref=administrator",
-		"javax.portlet.supports.mime-type=text/html"
-	},
-	service = Portlet.class
-)
-public class UsersAdminPortlet extends MVCPortlet {
-}
