--- modules/apps/foundation/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/portlet/configuration/icon/EditOrganizationPortletConfigurationIcon.java
@@ -1,120 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.users.admin.web.portlet.configuration.icon;
-
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.model.Organization;
-import com.liferay.portal.kernel.portlet.PortletURLFactoryUtil;
-import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIcon;
-import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.service.permission.OrganizationPermissionUtil;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.users.admin.constants.UsersAdminPortletKeys;
-import com.liferay.users.admin.web.portlet.action.ActionUtil;
-
-import javax.portlet.PortletRequest;
-import javax.portlet.PortletResponse;
-import javax.portlet.PortletURL;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Pei-Jung Lan
- */
-@Component(
-	immediate = true,
-	property = {
-		"javax.portlet.name=" + UsersAdminPortletKeys.USERS_ADMIN,
-		"path=/users_admin/view"
-	},
-	service = PortletConfigurationIcon.class
-)
-public class EditOrganizationPortletConfigurationIcon
-	extends BasePortletConfigurationIcon {
-
-	@Override
-	public String getMessage(PortletRequest portletRequest) {
-		return LanguageUtil.get(
-			getResourceBundle(getLocale(portletRequest)), "edit");
-	}
-
-	@Override
-	public String getURL(
-		PortletRequest portletRequest, PortletResponse portletResponse) {
-
-		try {
-			PortletURL portletURL = PortletURLFactoryUtil.create(
-				portletRequest, UsersAdminPortletKeys.USERS_ADMIN,
-				PortletRequest.RENDER_PHASE);
-
-			portletURL.setParameter(
-				"mvcRenderCommandName", "/users_admin/edit_organization");
-			portletURL.setParameter(
-				"redirect", _portal.getCurrentURL(portletRequest));
-
-			Organization organization = ActionUtil.getOrganization(
-				portletRequest);
-
-			portletURL.setParameter(
-				"organizationId",
-				String.valueOf(organization.getOrganizationId()));
-
-			return portletURL.toString();
-		}
-		catch (Exception e) {
-		}
-
-		return StringPool.BLANK;
-	}
-
-	@Override
-	public double getWeight() {
-		return 105;
-	}
-
-	@Override
-	public boolean isShow(PortletRequest portletRequest) {
-		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		try {
-			PermissionChecker permissionChecker =
-				themeDisplay.getPermissionChecker();
-
-			Organization organization = ActionUtil.getOrganization(
-				portletRequest);
-
-			if (OrganizationPermissionUtil.contains(
-					permissionChecker, organization, ActionKeys.UPDATE)) {
-
-				return true;
-			}
-		}
-		catch (Exception e) {
-		}
-
-		return false;
-	}
-
-	@Reference
-	private Portal _portal;
-
-}
