--- modules/apps/foundation/users-admin/users-admin-web/src/main/resources/com/liferay/users/admin/web/portlet/action/packageinfo
@@ -1,1 +1,0 @@
-version 2.3.0
