--- modules/apps/foundation/users-admin/users-admin-web/src/main/resources/com/liferay/users/admin/web/portlet/configuration/icon/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.1
